move ShareLinkSetting
