+++ src/client/js/components/ShareLink/ShareLink.jsx
@@ -7,7 +7,7 @@ import { withUnstatedContainers } from '../UnstatedUtils';
 import AppContainer from '../../services/AppContainer';
 import PageContainer from '../../services/PageContainer';
 
-import ShareLinkList from '../ShareLinkList';
+import ShareLinkList from './ShareLinkList';
 import ShareLinkForm from './ShareLinkForm';
 
 import { toastSuccess, toastError } from '../../util/apiNotification';
